Make services more composable
Additionally:
- Hard-code client used by service wrappers (as suggested by Tito Costa)
- Remove config class for S3 for consistency
- Convert ET and Rekognition into singletons
